backend/requirements.txt
google-genai==1.9.0
google-api-core==2.24.2
